Added Redis backup and restore capabilities

+++ src/nova/redis/RedisApp.cc
@@ -47,8 +47,10 @@ using std::stringstream;
 namespace nova { namespace redis {
 
 RedisApp::RedisApp(RedisAppStatusPtr status,
+                   nova::backup::BackupRestoreManagerPtr backup_restore_manager,
                    const int state_change_wait_time)
-:   DatastoreApp("redis-server", status, state_change_wait_time)
+:   DatastoreApp("redis-server", status, state_change_wait_time),
+    backup_restore_manager(backup_restore_manager)
 {
 }
 
@@ -96,24 +98,35 @@ void RedisApp::prepare(const optional<string> & json_root_password,
                        const optional<string> & overrides,
                        optional<BackupRestoreInfo> restore)
 {
-    if (!json_root_password) {
-        NOVA_LOG_ERROR("Missing root password!");
-        throw RedisException(RedisException::MISSING_ROOT_PASSWORD);
+    string root_password;
+    if (restore) {
+        if (status->is_running()) {
+            NOVA_LOG_INFO("Stopping redis instance before extracting backup.");
+            internal_stop(false);
+        }
+        NOVA_LOG_INFO("A restore was requested. Running now...");
+        backup_restore_manager->run(restore.get());
+        NOVA_LOG_INFO("Finished extracting backup.");
+        NOVA_LOG_INFO("Resetting config (but keeping password).");
+        reset_configuration(config_contents);
+    } else {
+        if (!json_root_password) {
+            NOVA_LOG_ERROR("Missing root password!");
+            throw RedisException(RedisException::MISSING_ROOT_PASSWORD);
+        }
+        NOVA_LOG_INFO("Writing Redis config file.");
+        root_password = json_root_password.get();
+        SHELL("sudo rm /etc/redis/redis.conf");
+        SHELL("sudo chmod -R 777 /etc/redis");
+        write_config(config_contents, root_password);
     }
-    const string root_password = json_root_password.get();
-
-    SHELL("sudo rm /etc/redis/redis.conf");
-    SHELL("sudo chmod -R 777 /etc/redis");
-
-    write_config(config_contents, root_password);
-
 
     NOVA_LOG_INFO("Starting to muck with Redis's config file.");
-    SHELL("mkdir -p /etc/redis/conf.d");
+    SHELL("sudo mkdir -p /etc/redis/conf.d");
+    SHELL("sudo chown -R nova:nova /etc/redis/conf.d");
+
     NOVA_LOG_INFO("Checking for Redis file.");
-    if (is_file("/etc/redis/conf.d/local.conf")) {
-        SHELL("rm /etc/redis/conf.d/local.conf");
-    }
+    SHELL("sudo rm -f /etc/redis/conf.d/local.conf");
 
     NOVA_LOG_INFO("Opening /etc/redis/conf.d/local.conf for writing");
     {
